style: factor out breadcrumb hook + fix no-title sidebar
