+++ src/js/components/Scope/ScopedTitle.tsx
@@ -2,25 +2,18 @@ import { useMemo, useState } from 'react';
 import { Link, useLocation } from 'react-router-dom';
 
 import { Breadcrumb, type BreadcrumbProps, Button, Flex, Space, Tooltip } from 'antd';
-import { ArrowLeftOutlined, HomeOutlined, ProfileOutlined, QuestionOutlined } from '@ant-design/icons';
+import { ArrowLeftOutlined, ProfileOutlined, QuestionOutlined } from '@ant-design/icons';
 import type { BreadcrumbItemType } from 'antd/es/breadcrumb/Breadcrumb';
 import CurrentPageHelpModal from '@/components/Util/CurrentPageHelpModal';
 import ScopePickerModal from './ScopePickerModal';
 
-import { useSelectedScope, useSelectedScopeTitles } from '@/features/metadata/hooks';
+import { useSelectedScope } from '@/features/metadata/hooks';
 import { useSearchQueryParams } from '@/features/search/hooks';
-import { useExtraBreadcrumb } from '@/features/ui/hooks';
-import { useLanguage, useTranslationFn } from '@/hooks';
-import {
-  useGetRouteTitleAndIcon,
-  useNavigateToRoot,
-  useNavigateToSameScopeUrl,
-  useNavigateToScope,
-} from '@/hooks/navigation';
+import { useTranslationFn } from '@/hooks';
+import { useNavigateToRoot, useNavigateToSameScopeUrl, useNavigateToScope } from '@/hooks/navigation';
 import { BentoRoute, TOP_LEVEL_ONLY_ROUTES } from '@/types/routes';
 import { getCurrentPage } from '@/utils/router';
 import { buildQueryParamsUrl } from '@/features/search/utils';
-import { useSmallScreen } from '@/hooks/useResponsiveContext';
 
 const NO_BACK_BUTTON = [undefined, undefined] as const;
 
@@ -76,16 +69,11 @@ const breadcrumbRender: BreadcrumbProps['itemRender'] = (route, _params, routes,
   return isLast || !route.path ? <span>{route.title}</span> : <Link to={{ pathname: route.path }}>{route.title}</Link>;
 };
 
-const ScopedTitle = () => {
+const ScopedTitle = ({ breadcrumbItems }: { breadcrumbItems: BreadcrumbItemType[] }) => {
   const location = useLocation();
   const t = useTranslationFn();
-  const language = useLanguage();
-  const isSmallScreen = useSmallScreen();
 
-  const { scope, fixedProject, fixedDataset } = useSelectedScope();
-  const { projectTitle, datasetTitle } = useSelectedScopeTitles();
-
-  const getRouteTitleAndIcon = useGetRouteTitleAndIcon();
+  const { fixedProject, fixedDataset } = useSelectedScope();
 
   const currentPage = getCurrentPage();
   const scopeSelectionEnabled =
@@ -107,63 +95,6 @@ const ScopedTitle = () => {
     const k = `page_help.${currentPage}`;
     return k !== t(k);
   }, [t, currentPage]);
-
-  const extraBreadcrumb = useExtraBreadcrumb();
-
-  const breadcrumbItems: BreadcrumbItemType[] = useMemo(() => {
-    const currentPageTitle = t(getRouteTitleAndIcon(currentPage)[0]);
-
-    const items: BreadcrumbItemType[] = [];
-
-    if (scope.project && !fixedProject) {
-      // If we have more than one project (or we're in catalogue mode), fixedProject will be false, meaning we should
-      // show project context in the navigation:
-      items.push({
-        title: projectTitle,
-        path: `/${language}/p/${scope.project}`,
-      });
-    }
-
-    if (scope.dataset && !fixedDataset) {
-      // If we have a dataset selected, and we don't have just a single project+dataset, we should show the dataset
-      // context in the navigation.
-      if (scope.project && fixedProject) {
-        // If we additionally have a fixed project, we can "anchor" the dataset visually vs. the root page (which isn't
-        // the catalogue, but rather the project) using a home icon:
-        items.push({
-          title: <HomeOutlined />,
-          path: `/${language}/`,
-        });
-      }
-      items.push({
-        title: datasetTitle,
-        path: `/${language}/d/${scope.dataset}`,
-      });
-    }
-
-    // We treat the overview as the "default" page, meaning we won't show the page name in the breadcrumb bar.
-    if (currentPage !== BentoRoute.Overview) {
-      items.push({ title: currentPageTitle });
-    }
-
-    if (extraBreadcrumb) {
-      items.push(extraBreadcrumb);
-    }
-
-    return isSmallScreen && items.length ? [items.at(-1)!] : items;
-  }, [
-    language,
-    t,
-    projectTitle,
-    datasetTitle,
-    scope,
-    fixedProject,
-    fixedDataset,
-    currentPage,
-    getRouteTitleAndIcon,
-    extraBreadcrumb,
-    isSmallScreen,
-  ]);
 
   const [backClickText, onBackClick] = useBackButtonInfo();
